requirements.txt
Django==1.10.1
django-filter==0.14.0
djangorestframework==3.4.6
djangorestframework-jwt==1.8.0
Markdown==2.6.6
MySQL-python==1.2.5
PyJWT==1.4.2
